Finish for publish
* Fix all methods to compiling

* Finish all details to publish

+++ src/xtea.cpp
@@ -22,7 +22,7 @@ namespace xtea {
 	}
 
 	void encrypt(uint8_t* data, size_t length, const round_keys& k) {
-		for (int32_t i = 0; i < k.size(); i += 2) {
+		for (int32_t i = 0; i < static_cast<int32_t>(k.size()); i += 2) {
 			for (auto it = data, last = data + length; it < last; it += 8) {
 				uint32_t left, right;
 				std::memcpy(&left, it, 4);
